requirements.txt
discord.py
python-dotenv

# Testing dependencies
pytest
pytest-asyncio
pytest-mock